Add root dir conftest file
