+++ openfe/tests/conftest.py
@@ -11,28 +11,6 @@ import openfe
 from gufe import SmallMoleculeComponent, LigandAtomMapping
 
 
-# allow for optional slow tests
-# See: https://docs.pytest.org/en/latest/example/simple.html
-def pytest_addoption(parser):
-    parser.addoption(
-        "--runslow", action="store_true", default=False, help="run slow tests"
-    )
-
-
-def pytest_configure(config):
-    config.addinivalue_line("markers", "slow: mark test as slow to run")
-
-
-def pytest_collection_modifyitems(config, items):
-    if config.getoption("--runslow"):
-        # --runslow given in cli: do not skip slow tests
-        return
-    skip_slow = pytest.mark.skip(reason="need --runslow to run")
-    for item in items:
-        if "slow" in item.keywords:
-            item.add_marker(skip_slow)
-
-
 def mol_from_smiles(smiles: str) -> Chem.Mol:
     m = Chem.MolFromSmiles(smiles)
     AllChem.Compute2DCoords(m)
